third_party/blink/web_tests/external/wpt/html/browsers/history/the-location-interface/assign_before_load-expected.txt

This is a testharness.js-based test.
FAIL Assignment to location before document is completely loaded assert_equals: expected 2 but got 1
Harness: the test ran to completion.
